+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentMethodType.swift
@@ -17,14 +17,6 @@ extension PaymentSheet {
 
         func supportsAddingRequirements() -> [PaymentMethodTypeRequirement] {
             switch self {
-            case .dynamic("revolut_pay"):
-                return [.returnURL]
-            case .dynamic("amazon_pay"):
-                return [.returnURL]
-            case .dynamic("mobilepay"):
-                return [.returnURL]
-            case .dynamic("zip"):
-                return [.returnURL]
             default:
                 return [.unsupported]
             }
@@ -91,14 +83,6 @@ extension PaymentSheet {
         var displayName: String {
             if let stpPaymentMethodType = stpPaymentMethodType {
                 return stpPaymentMethodType.displayName
-            } else if case .dynamic("revolut_pay") = self {
-                return "Revolut Pay"
-            } else if case .dynamic("mobilepay") = self {
-                return "MobilePay"
-            } else if case .dynamic("zip") = self {
-                return "Zip"
-            } else if case .dynamic("amazon_pay") = self {
-                return "Amazon Pay"
             } else if case .dynamic(let name) = self {
                 // TODO: We should introduce a display name in our model rather than presenting the payment method type
                 return name
@@ -336,20 +320,20 @@ extension PaymentSheet {
                     switch stpPaymentMethodType {
                     case .card:
                         return []
-                    case .alipay, .payPal, .cashApp:
+                    case .payPal, .cashApp, .revolutPay:
                         return [.returnURL]
-                    case .USBankAccount:
+                    case .USBankAccount, .boleto:
                         return [.userSupportsDelayedPaymentMethods]
                     case .sofort, .iDEAL, .bancontact:
                         // n.b. While sofort, iDEAL, and bancontact are themselves not delayed, they turn into SEPA upon save, which IS delayed.
                         return [.returnURL, .userSupportsDelayedPaymentMethods]
-                    case .SEPADebit:
+                    case .SEPADebit, .AUBECSDebit:
                         return [.userSupportsDelayedPaymentMethods]
                     case .bacsDebit:
                         return [.returnURL, .userSupportsDelayedPaymentMethods]
-                    case .AUBECSDebit, .cardPresent, .blik, .weChatPay, .grabPay, .FPX, .giropay, .przelewy24, .EPS,
-                        .netBanking, .OXXO, .afterpayClearpay, .UPI, .boleto, .klarna, .link, .linkInstantDebit,
-                        .affirm, .unknown, .paynow:
+                    case .cardPresent, .blik, .weChatPay, .grabPay, .FPX, .giropay, .przelewy24, .EPS,
+                        .netBanking, .OXXO, .afterpayClearpay, .UPI, .klarna, .link, .linkInstantDebit,
+                        .affirm, .paynow, .zip, .amazonPay, .alma, .mobilePay, .unknown, .alipay, .konbini, .promptPay:
                         return [.unsupportedForSetup]
                     @unknown default:
                         return [.unsupportedForSetup]
@@ -358,17 +342,17 @@ extension PaymentSheet {
             } else {
                 requirements = {
                     switch stpPaymentMethodType {
-                    case .blik, .card, .cardPresent, .UPI, .weChatPay, .paynow:
+                    case .blik, .card, .cardPresent, .UPI, .weChatPay, .paynow, .promptPay:
                         return []
                     case .alipay, .EPS, .FPX, .giropay, .grabPay, .netBanking, .payPal, .przelewy24, .klarna,
-                            .linkInstantDebit, .bancontact, .iDEAL, .cashApp, .affirm:
+                            .linkInstantDebit, .bancontact, .iDEAL, .cashApp, .affirm, .zip, .revolutPay, .amazonPay, .alma, .mobilePay:
                         return [.returnURL]
                     case .USBankAccount:
                         return [
                             .userSupportsDelayedPaymentMethods, .financialConnectionsSDK,
                             .validUSBankVerificationMethod,
                         ]
-                    case .OXXO, .boleto, .AUBECSDebit, .SEPADebit:
+                    case .OXXO, .boleto, .AUBECSDebit, .SEPADebit, .konbini:
                         return [.userSupportsDelayedPaymentMethods]
                     case .bacsDebit, .sofort:
                         return [.returnURL, .userSupportsDelayedPaymentMethods]