+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetFormFactory/PaymentSheetFormFactory.swift
@@ -146,21 +146,28 @@ class PaymentSheetFormFactory {
         } else if paymentMethod == .USBankAccount {
             return makeUSBankAccount(merchantName: configuration.merchantDisplayName)
         } else if paymentMethod == .UPI {
-            return makeDefaultsApplierWrapper(for: makeUPI())
+            return makeUPI()
         } else if paymentMethod == .cashApp && saveMode == .merchantRequired {
             // special case, display mandate for Cash App when setting up or pi+sfu
             additionalElements = [makeCashAppMandate()]
         } else if paymentMethod.stpPaymentMethodType == .payPal && saveMode == .merchantRequired {
             // Paypal requires mandate when setting up
             additionalElements = [makePaypalMandate()]
+        } else if paymentMethod.stpPaymentMethodType == .revolutPay && saveMode == .merchantRequired {
+            // special case, display mandate for revolutPay when setting up or pi+sfu
+            additionalElements = [makeRevolutPayMandate()]
         } else if paymentMethod.stpPaymentMethodType == .bancontact {
             return makeBancontact()
         } else if paymentMethod.stpPaymentMethodType == .blik {
-            return makeDefaultsApplierWrapper(for: makeBLIK())
+            return makeBLIK()
         } else if paymentMethod == .externalPayPal {
             return makeExternalPayPal()
         } else if paymentMethod.stpPaymentMethodType == .OXXO {
-            return makeDefaultsApplierWrapper(for: makeOXXO())
+            return  makeOXXO()
+        } else if paymentMethod.stpPaymentMethodType == .konbini {
+            return makeKonbini()
+        } else if paymentMethod.stpPaymentMethodType == .boleto {
+            return makeBoleto()
         }
 
         guard let spec = specFromJSONProvider() else {
@@ -295,6 +302,11 @@ extension PaymentSheetFormFactory {
 
     func makeCashAppMandate() -> PaymentMethodElement {
         let mandateText = String(format: String.Localized.cash_app_mandate_text, configuration.merchantDisplayName, configuration.merchantDisplayName)
+        return makeMandate(mandateText: mandateText)
+    }
+
+    func makeRevolutPayMandate() -> PaymentMethodElement {
+        let mandateText = String(format: String.Localized.revolut_pay_mandate_text, configuration.merchantDisplayName)
         return makeMandate(mandateText: mandateText)
     }
 
@@ -504,6 +516,18 @@ extension PaymentSheetFormFactory {
             merchantName: merchantName,
             theme: theme
         )
+    }
+
+    func makeKonbini() -> PaymentMethodElement {
+        let contactInfoSection = makeContactInformationSection(nameRequiredByPaymentMethod: true, emailRequiredByPaymentMethod: true, phoneRequiredByPaymentMethod: false)
+        let billingDetails = makeBillingAddressSectionIfNecessary(requiredByPaymentMethod: false)
+        let konbiniPhoneNumber = PaymentMethodElementWrapper(TextFieldElement.makeKonbini(theme: theme)) { textField, params in
+            params.confirmPaymentMethodOptions.konbiniOptions = .init()
+            params.confirmPaymentMethodOptions.konbiniOptions?.confirmationNumber = textField.text
+            return params
+        }
+        let elements = [contactInfoSection, konbiniPhoneNumber, billingDetails].compactMap { $0 }
+        return FormElement(autoSectioningElements: elements, theme: theme)
     }
 
     func makeExternalPayPal() -> PaymentMethodElement {