+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/ViewControllers/PollingViewModel.swift
@@ -15,13 +15,15 @@ import UIKit
 class PollingViewModel {
 
     let paymentMethodType: STPPaymentMethodType
-    let supportedPaymentMethods: [STPPaymentMethodType] = [.UPI, .blik]
+    let supportedPaymentMethods: [STPPaymentMethodType] = [.UPI, .blik, .paynow, .promptPay]
     lazy var CTA: String = {
         switch paymentMethodType {
         case .UPI:
             return .Localized.open_upi_app
         case .blik:
             return .Localized.blik_confirm_payment
+        case .paynow, .promptPay:
+            return .Localized.paynow_confirm_payment
         default:
             fatalError("Polling CTA has not been implemented for \(paymentMethodType)")
         }
@@ -32,10 +34,22 @@ class PollingViewModel {
             return Date().addingTimeInterval(60 * 5) // 5 minutes
         case .blik:
             return Date().addingTimeInterval(60) // 60 seconds
+        case .paynow, .promptPay:
+            return Date().addingTimeInterval(60 * 60) // 1 hour
         default:
             fatalError("Polling deadline has not been implemented for \(paymentMethodType)")
         }
     }()
+    var retryInterval: TimeInterval {
+        switch paymentMethodType {
+        case .blik, .paynow, .promptPay:
+            return 1
+        case .UPI:
+            return 10
+        default:
+            fatalError("Polling retry interval has not been implemented for \(paymentMethodType)")
+        }
+    }
 
     init(paymentMethodType: STPPaymentMethodType) {
         guard supportedPaymentMethods.contains(paymentMethodType) else {