+++ StripePayments/StripePayments/Source/PaymentHandler/STPPaymentHandler.swift
@@ -605,7 +605,7 @@ public class STPPaymentHandler: NSObject {
     /// because the funds can take up to 14 days to transfer from the customer's bank.
     class func _isProcessingIntentSuccess(for type: STPPaymentMethodType) -> Bool {
         switch type {
-        // Asynchronous
+        // Asynchronous payment methods whose intent.status is 'processing' after handling the next action
         case .SEPADebit,
             .bacsDebit,  // Bacs Debit takes 2-3 business days
             .AUBECSDebit,
@@ -637,7 +637,14 @@ public class STPPaymentHandler: NSObject {
             .affirm,
             .linkInstantDebit,
             .cashApp,
-            .paynow:
+            .paynow,
+            .zip,
+            .revolutPay,
+            .mobilePay,
+            .amazonPay,
+            .alma,
+            .konbini,
+            .promptPay:
             return false
 
         case .unknown:
@@ -648,7 +655,7 @@ public class STPPaymentHandler: NSObject {
         }
     }
 
-            func _handleNextAction(
+    func _handleNextAction(
         forPayment paymentIntent: STPPaymentIntent,
         with authenticationContext: STPAuthenticationContext,
         returnURL returnURLString: String?,
@@ -1284,7 +1291,7 @@ public class STPPaymentHandler: NSObject {
                 currentAction.complete(with: .succeeded, error: nil)
                 return
             }
-            presentingVC.presentPollingVCForAction(action: currentAction, type: .blik)
+            presentingVC.presentPollingVCForAction(action: currentAction, type: .blik, safariViewController: nil)
 
         case .verifyWithMicrodeposits:
             // The customer must authorize after the microdeposits appear in their bank account
@@ -1299,7 +1306,7 @@ public class STPPaymentHandler: NSObject {
                 return
             }
 
-            presentingVC.presentPollingVCForAction(action: currentAction, type: .UPI)
+            presentingVC.presentPollingVCForAction(action: currentAction, type: .UPI, safariViewController: nil)
         case .cashAppRedirectToApp:
             guard
                 let returnURL = URL(string: currentAction.returnURLString ?? "")
@@ -1319,6 +1326,50 @@ public class STPPaymentHandler: NSObject {
                         ]
                     )
                 )
+            }
+        case .payNowDisplayQrCode:
+            guard
+                let returnURL = URL(string: currentAction.returnURLString ?? ""),
+                let presentingVC = currentAction.authenticationContext
+                    as? PaymentSheetAuthenticationContext,
+                let hostedInstructionsURL = authenticationAction.payNowDisplayQrCode?.hostedInstructionsURL
+
+            else {
+                fatalError()
+            }
+
+            _handleRedirect(to: hostedInstructionsURL, fallbackURL: hostedInstructionsURL, return: returnURL) { safariViewController in
+                // Present the polling view controller behind the web view so we can start polling right away
+                presentingVC.presentPollingVCForAction(action: currentAction, type: .paynow, safariViewController: safariViewController)
+            }
+        case .konbiniDisplayDetails:
+            if let hostedVoucherURL = authenticationAction.konbiniDisplayDetails?.hostedVoucherURL {
+                self._handleRedirect(to: hostedVoucherURL, withReturn: nil)
+            } else {
+                currentAction.complete(
+                    with: STPPaymentHandlerActionStatus.failed,
+                    error: _error(
+                        for: .unsupportedAuthenticationErrorCode,
+                        userInfo: [
+                            "STPIntentAction": authenticationAction.description,
+                        ]
+                    )
+                )
+            }
+        case .promptpayDisplayQrCode:
+            guard
+                let returnURL = URL(string: currentAction.returnURLString ?? ""),
+                let presentingVC = currentAction.authenticationContext
+                    as? PaymentSheetAuthenticationContext,
+                let hostedInstructionsURL = authenticationAction.promptPayDisplayQrCode?.hostedInstructionsURL
+
+            else {
+                fatalError()
+            }
+
+            _handleRedirect(to: hostedInstructionsURL, fallbackURL: hostedInstructionsURL, return: returnURL) { safariViewController in
+                // Present the polling view controller behind the web view so we can start polling right away
+                presentingVC.presentPollingVCForAction(action: currentAction, type: .promptPay, safariViewController: safariViewController)
             }
         @unknown default:
             fatalError()
@@ -1444,7 +1495,10 @@ public class STPPaymentHandler: NSObject {
                                                     retryCount: retryCount - 1
                                                 )
                                             }
-                                        } else {
+                                        } else if retrievedPaymentIntent?.paymentMethod?.type != .paynow
+                                                    && retrievedPaymentIntent?.paymentMethod?.type != .promptPay {
+                                            // For PayNow, we don't want to mark as canceled when the web view dismisses
+                                            // Instead we rely on the presented PollingViewController to complete the currentAction
                                             self._markChallengeCanceled(withCompletion: { _, _ in
                                                 // We don't forward cancelation errors
                                                 currentAction.complete(
@@ -1581,13 +1635,24 @@ public class STPPaymentHandler: NSObject {
         )
     }
 
-    /// This method:
-    /// 1. Redirects to an app using url
-    /// 2. Open fallbackURL in a webview if 1) fails
-            ///
-    func _handleRedirect(to nativeURL: URL?, fallbackURL: URL?, return returnURL: URL?) {
+    /// Handles redirection to URLs using a native URL or a fallback URL and updates the current action.
+    /// Redirects to an app if possible, if that fails opens the url in a web view
+    /// - Parameters:
+    ///     - nativeURL: A URL to be opened natively.
+    ///     - fallbackURL: A secondary URL to be attempted if the native URL is not available.
+    ///     - returnURL: The URL to be registered with the `STPURLCallbackHandler`.
+    ///     - completion: A completion block invoked after the URL redirection is handled. The SFSafariViewController used is provided as an argument, if it was used for the redirect.
+    func _handleRedirect(to nativeURL: URL?, fallbackURL: URL?, return returnURL: URL?, completion: ((SFSafariViewController?) -> Void)? = nil) {
         if let redirectShim = _redirectShim, let url = nativeURL ?? fallbackURL {
             redirectShim(url, returnURL, true)
+        }
+
+        // During testing, the completion block is not called since the `UIApplication.open` completion block is never invoked.
+        // As a workaround we invoke the completion in a defer block if the _redirectShim is not nil to simulate presenting a web view
+        defer {
+            if _redirectShim != nil {
+                completion?(nil)
+            }
         }
 
         var url = nativeURL
@@ -1632,7 +1697,9 @@ public class STPPaymentHandler: NSObject {
                     }
                     safariViewController.delegate = self
                     self.safariViewController = safariViewController
-                    presentingViewController.present(safariViewController, animated: true)
+                    presentingViewController.present(safariViewController, animated: true, completion: {
+                      completion?(safariViewController)
+                    })
                 } else {
                     currentAction.complete(
                         with: STPPaymentHandlerActionStatus.failed,
@@ -1679,6 +1746,7 @@ public class STPPaymentHandler: NSObject {
                         // no app installed, launch safari view controller
                         presentSFViewControllerBlock()
                     } else {
+                        completion?(nil)
                         NotificationCenter.default.addObserver(
                             self,
                             selector: #selector(self._handleWillForegroundNotification),
@@ -1751,10 +1819,13 @@ public class STPPaymentHandler: NSObject {
                 .useStripeSDK,
                 .alipayHandleRedirect,
                 .weChatPayRedirectToApp,
-                .cashAppRedirectToApp:
+                .cashAppRedirectToApp,
+                .payNowDisplayQrCode,
+                .promptpayDisplayQrCode:
                 return false
             case .OXXODisplayDetails,
                 .boletoDisplayDetails,
+                .konbiniDisplayDetails,
                 .verifyWithMicrodeposits,
                 .BLIKAuthorize,
                 .upiAwaitNotification:
@@ -1782,7 +1853,7 @@ public class STPPaymentHandler: NSObject {
             threeDSSourceID = nextAction.useStripeSDK?.threeDSSourceID
         case .OXXODisplayDetails, .alipayHandleRedirect, .unknown, .BLIKAuthorize,
             .weChatPayRedirectToApp, .boletoDisplayDetails, .verifyWithMicrodeposits,
-            .upiAwaitNotification, .cashAppRedirectToApp:
+            .upiAwaitNotification, .cashAppRedirectToApp, .konbiniDisplayDetails, .payNowDisplayQrCode, .promptpayDisplayQrCode:
             break
         @unknown default:
             fatalError()
@@ -2260,7 +2331,7 @@ extension STPPaymentHandler {
         if let paymentSheet = currentAction.authenticationContext
             .authenticationPresentingViewController() as? PaymentSheetAuthenticationContext
         {
-            paymentSheet.dismiss(challengeViewController)
+            paymentSheet.dismiss(challengeViewController, completion: nil)
         } else {
             challengeViewController.dismiss(animated: true, completion: nil)
         }
@@ -2278,8 +2349,8 @@ extension STPPaymentHandler {
 /// Internal authentication context for PaymentSheet magic
 @_spi(STP) public protocol PaymentSheetAuthenticationContext: STPAuthenticationContext {
     func present(_ authenticationViewController: UIViewController, completion: @escaping () -> Void)
-    func dismiss(_ authenticationViewController: UIViewController)
-    func presentPollingVCForAction(action: STPPaymentHandlerActionParams, type: STPPaymentMethodType)
+    func dismiss(_ authenticationViewController: UIViewController, completion: (() -> Void)?)
+    func presentPollingVCForAction(action: STPPaymentHandlerActionParams, type: STPPaymentMethodType, safariViewController: SFSafariViewController?)
 }
 
 @_spi(STP) public protocol FormSpecPaymentHandler {